MAGENTO2-406 - Move stocklogic to own handler
